+++ entities/build.gradle
@@ -1,0 +1,9 @@
+apply plugin: 'java'
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.google.code.gson:gson:2.7'
+}
+
+sourceCompatibility = "1.7"
+targetCompatibility = "1.7"
