+++ entities/src/main/java/com/tapglue/android/entities/Comment.java
