+++ entities/src/main/java/com/tapglue/android/entities/Connection.java
