+++ entities/src/main/java/com/tapglue/android/entities/ConnectionList.java
