+++ entities/src/main/java/com/tapglue/android/entities/Event.java
