+++ entities/src/main/java/com/tapglue/android/entities/Follow.java
