+++ entities/src/main/java/com/tapglue/android/entities/Friend.java
