+++ entities/src/main/java/com/tapglue/android/entities/Like.java
