+++ entities/src/main/java/com/tapglue/android/entities/NewsFeed.java
