+++ entities/src/main/java/com/tapglue/android/entities/Post.java
