+++ entities/src/main/java/com/tapglue/android/entities/Reaction.java
