+++ entities/src/main/java/com/tapglue/android/http/ApiPage.java
