+++ entities/src/main/java/com/tapglue/android/http/CommentsFeed.java
