+++ entities/src/main/java/com/tapglue/android/http/ErrorFeed.java
