+++ entities/src/main/java/com/tapglue/android/http/FlattenableFeed.java
