+++ entities/src/main/java/com/tapglue/android/http/TapglueError.java
