+++ entities/src/main/java/com/tapglue/android/http/payloads/EmailLoginPayload.java
