+++ entities/src/main/java/com/tapglue/android/http/payloads/SocialSearchPayload.java
