+++ entities/src/main/java/com/tapglue/android/http/payloads/UsernameLoginPayload.java
