+++ tapglue-android-sdk/build.gradle
@@ -87,6 +87,7 @@ dependencies {
     compile 'com.squareup.retrofit2:adapter-rxjava2:2.2.0'
     compile 'com.squareup.retrofit2:converter-gson:2.2.0'
     compile 'com.google.code.gson:gson:2.7'
+    compile project(':entities')
     compile 'com.squareup.okhttp3:logging-interceptor:3.3.1'
     compile 'com.android.support:appcompat-v7:23.2.1'
     compile 'com.google.firebase:firebase-messaging:9.0.2'
